Removing Focal Loss, and adding it to separate .py file
