+++ run_bart_detection.sh
